+++ CHANGELOG.md
@@ -15,5 +15,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * `hx_push_url` and `hx_push_url_str`
 * `hx_swap_oob` and `hx_swap_oob_swap`
 * `hx_boost`
+* `hx_on`, `hx_on_htmx_before_request` and `hx_on_htmx_after_request`
 
 [Unreleased]: https://github.com/jcornaz/fun-htmx/compare/...HEAD